Use ITF8 in DistributedCacheExtensions
